Fix "back" button preventing further input and use my first Ref :')
Fixes this problem:

1. start at / (maybe refresh for good measure)
2. enter any code and hit submit (one of the examples will do)
3. should see twinkling code, fine
4. go back in your browser
5. try to enter new code, or click another example
5b. ?? textarea doesn't change

It was because App and Form both had a 'userCode' state. I'm not sure I intended that... but anyway, this is tidier and I think probably faster too. No onChange needed on the textarea.

Does mean we've lost some control over the textarea (I think? I'm new to useRef), but we weren't using that.

https://reactjs.org/docs/hooks-reference.html#useref
^ notes that a 'callback ref' might be useful if we want to listen harder to it in future.